requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.11
django-allauth==0.54.0
gunicorn==20.1.0
idna==3.4
oauthlib==3.2.2
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.3
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.5.0
Unipath==1.1
urllib3==1.26.15
whitenoise==5.3.0
